+++ src/mruby_integration/core/drawing.cpp
@@ -1,7 +1,6 @@
 #include "mruby.h"
 #include "raylib.h"
 
-#include "mruby_integration/models/colour.hpp"
 #include "mruby_integration/struct_types.hpp"
 
 #include "ruby/core/drawing.hpp"
@@ -13,20 +12,6 @@ mrb_clear_background(mrb_state* mrb, mrb_value) -> mrb_value
   mrb_get_args(mrb, "d", &colour, &Color_type);
 
   ClearBackground(*colour);
-  return mrb_nil_value();
-}
-
-auto
-mrb_begin_drawing(mrb_state*, mrb_value) -> mrb_value
-{
-  BeginDrawing();
-  return mrb_nil_value();
-}
-
-auto
-mrb_end_drawing(mrb_state*, mrb_value) -> mrb_value
-{
-  EndDrawing();
   return mrb_nil_value();
 }
 
@@ -106,13 +91,6 @@ append_core_drawing(mrb_state* mrb)
                     "clear_background",
                     mrb_clear_background,
                     MRB_ARGS_REQ(1));
-  mrb_define_method(mrb,
-                    mrb->kernel_module,
-                    "begin_drawing",
-                    mrb_begin_drawing,
-                    MRB_ARGS_NONE());
-  mrb_define_method(
-    mrb, mrb->kernel_module, "end_drawing", mrb_end_drawing, MRB_ARGS_NONE());
   mrb_define_method(mrb,
                     mrb->kernel_module,
                     "begin_texture_mode",