+++ src/mruby_integration/models/window.cpp
@@ -25,11 +25,11 @@ mrb_Window_open(mrb_state* mrb, mrb_value) -> mrb_value
   }
 
   // def self.open(width: 800, height: 480, title: "Taylor Game")
-  mrb_int kw_num = 3;
-  mrb_int kw_required = 0;
-  mrb_sym kw_names[] = { mrb_intern_lit(mrb, "width"),
-                         mrb_intern_lit(mrb, "height"),
-                         mrb_intern_lit(mrb, "title") };
+  const mrb_int kw_num = 3;
+  const mrb_int kw_required = 0;
+  const mrb_sym kw_names[] = { mrb_intern_lit(mrb, "width"),
+                               mrb_intern_lit(mrb, "height"),
+                               mrb_intern_lit(mrb, "title") };
   mrb_value kw_values[kw_num];
   mrb_kwargs kwargs = { kw_num, kw_required, kw_names, kw_values, nullptr };
   mrb_get_args(mrb, ":", &kwargs);
@@ -87,6 +87,20 @@ auto
 mrb_Window_ready(mrb_state*, mrb_value) -> mrb_value
 {
   return mrb_bool_value(IsWindowReady());
+}
+
+auto
+mrb_Window_begin_drawing(mrb_state*, mrb_value) -> mrb_value
+{
+  BeginDrawing();
+  return mrb_nil_value();
+}
+
+auto
+mrb_Window_end_drawing(mrb_state*, mrb_value) -> mrb_value
+{
+  EndDrawing();
+  return mrb_nil_value();
 }
 
 auto
@@ -374,6 +388,13 @@ append_models_Window(mrb_state* mrb)
     mrb, Window_class, "close", mrb_Window_close, MRB_ARGS_NONE());
   mrb_define_class_method(
     mrb, Window_class, "ready?", mrb_Window_ready, MRB_ARGS_NONE());
+  mrb_define_class_method(mrb,
+                          Window_class,
+                          "begin_drawing",
+                          mrb_Window_begin_drawing,
+                          MRB_ARGS_NONE());
+  mrb_define_class_method(
+    mrb, Window_class, "end_drawing", mrb_Window_end_drawing, MRB_ARGS_NONE());
   mrb_define_class_method(
     mrb, Window_class, "width", mrb_Window_width, MRB_ARGS_NONE());
   mrb_define_class_method(